Revert "Merge pull request #11 from couchbaselabs/couchbase3-dgm"
This reverts commit e3edaf454694b7e6cb42f20a9ebed77ac9fc05cb, reversing
changes made to 2b278c0e7a5d8aa81670612a1d948f8a275ec428.

+++ core/src/main/java/com/yahoo/ycsb/workloads/CoreWorkload.java
@@ -393,7 +393,7 @@ public class CoreWorkload extends Workload {
     recordcount =
         Long.parseLong(p.getProperty(Client.RECORD_COUNT_PROPERTY, Client.DEFAULT_RECORD_COUNT));
     if (recordcount == 0) {
-      recordcount = Long.MAX_VALUE;
+      recordcount = Integer.MAX_VALUE;
     }
     String requestdistrib =
         p.getProperty(REQUEST_DISTRIBUTION_PROPERTY, REQUEST_DISTRIBUTION_PROPERTY_DEFAULT);
@@ -409,7 +409,7 @@ public class CoreWorkload extends Workload {
     //System.err.println("printing INSERT_START_PROPERTY within " +
         //"Coreworkload init" + insertstart);
     long insertcount=
-        Long.parseLong(p.getProperty(INSERT_COUNT_PROPERTY, String.valueOf(recordcount - insertstart)));
+        Integer.parseInt(p.getProperty(INSERT_COUNT_PROPERTY, String.valueOf(recordcount - insertstart)));
     //System.err.println("printing INSERT_COUNT_PROPERTY within " +
         //"Coreworkload init" + insertstart);
     // Confirm valid values for insertstart and insertcount in relation to recordcount
@@ -471,8 +471,8 @@ public class CoreWorkload extends Workload {
       // the keyspace doesn't change from the perspective of the scrambled zipfian generator
       final double insertproportion = Double.parseDouble(
           p.getProperty(INSERT_PROPORTION_PROPERTY, INSERT_PROPORTION_PROPERTY_DEFAULT));
-      long opcount = Long.parseLong(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
-      long expectednewkeys = (long) ((opcount) * insertproportion * 2.0); // 2 is fudge factor
+      int opcount = Integer.parseInt(p.getProperty(Client.OPERATION_COUNT_PROPERTY));
+      int expectednewkeys = (int) ((opcount) * insertproportion * 2.0); // 2 is fudge factor
 
       keychooser = new ScrambledZipfianGenerator(insertstart, insertstart + insertcount + expectednewkeys);
     } else if (requestdistrib.compareTo("latest") == 0) {
@@ -582,7 +582,7 @@ public class CoreWorkload extends Workload {
    */
   @Override
   public boolean doInsert(DB db, Object threadstate) {
-    long keynum = keysequence.nextValue().longValue();
+    int keynum = keysequence.nextValue().intValue();
     String dbkey = buildKeyName(keynum);
 
     //System.err.println("keyname/dbkey value" + dbkey);
@@ -684,11 +684,11 @@ public class CoreWorkload extends Workload {
     long keynum;
     if (keychooser instanceof ExponentialGenerator) {
       do {
-        keynum = transactioninsertkeysequence.lastValue() - keychooser.nextValue().longValue();
+        keynum = transactioninsertkeysequence.lastValue() - keychooser.nextValue().intValue();
       } while (keynum < 0);
     } else {
       do {
-        keynum = keychooser.nextValue().longValue();
+        keynum = keychooser.nextValue().intValue();
       } while (keynum > transactioninsertkeysequence.lastValue());
     }
     return keynum;
@@ -775,7 +775,7 @@ public class CoreWorkload extends Workload {
     measurements.measureIntended("READ-MODIFY-WRITE", (int) ((en - ist) / 1000));
   }
 
-  public boolean doInsertCollection(DB db, Object threadstate, String scope, String collection, long keynum) {
+  public boolean doInsertCollection(DB db, Object threadstate, String scope, String collection, int keynum) {
     return false;
   }
