+++ core/src/main/java/com/yahoo/ycsb/workloads/CustomCollectionWorkload.java
@@ -656,9 +656,9 @@ public class CustomCollectionWorkload extends Workload {
   }
 
 
-  public boolean doInsertCollection(DB db, Object threadstate, String scope, String collection, long key) {
+  public boolean doInsertCollection(DB db, Object threadstate, String scope, String collection, int key) {
 
-    long keynum = key;
+    int keynum = key;
 
     String dbkey = buildKeyName(keynum);
